% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_h5ad.R
\name{from_h5ad}
\alias{from_h5ad}
\title{Copy trajectory from h5ad}
\usage{
from_h5ad(
  anndata,
  trajectory_prefix = "traj",
  dimred_prefix = trajectory_prefix,
  grouping_prefix = trajectory_prefix
)
}
\arguments{
\item{anndata}{AnnData object}

\item{name_prefix}{A prefix to add to each created object in the anndata.}
}
\description{
Copy trajectory from h5ad
}
