% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_h5ad.R
\name{to_h5ad}
\alias{to_h5ad}
\title{Copy trajectory to h5ad}
\usage{
to_h5ad(
  traj,
  anndata = NULL,
  trajectory_prefix = NULL,
  dimred_name = NULL,
  grouping_name = NULL,
  fimp_name = NULL
)
}
\arguments{
\item{traj}{Trajectory object}

\item{anndata}{AnnData object}

\item{trajectory_prefix}{The prefix of a trajectory to write.}

\item{dimred_name}{A dimred to write, a matrix in `ad$obsm`.}

\item{grouping_name}{A grouping to write, a column in `ad$obs`.}

\item{fimp_name}{A feature importance to write, a column in `ad$var`.}
}
\description{
Copy trajectory to h5ad
}
