% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_h5ad.R
\name{to_h5ad}
\alias{to_h5ad}
\title{Copy trajectory to h5ad}
\usage{
to_h5ad(anndata, traj, gimp = NULL, name_prefix = "")
}
\arguments{
\item{anndata}{AnnData object}

\item{traj}{Trajectory object}

\item{gimp}{Gene Importance object}

\item{name_prefix}{A prefix to add to each created object in the anndata.}
}
\description{
Copy trajectory to h5ad
}
